update movie review

+++ Backend/API Cinema Web/API-Cinema-Web/src/main/java/com/cinemaweb/API/Cinema/Web/entity/Movie.java
@@ -22,5 +22,5 @@ public class Movie {
     String movieGenre;
     int movieLength;  //this time watch film
     String movieDescription;
-    int movieReview;  //its star 1* 2* 3* ...
+    Double movieReview;  //its star 1* 2* 3* ...
 }
